+++ common/service/k8s/zpk/logic/helmmanifest_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/w7panel/w7panel/app/zpk/logic/types"
+	"github.com/w7panel/w7panel/common/service/k8s/zpk/logic/types"
 )
 
 func TestHelmManifestApp(t *testing.T) {